event refactor + string fix
