+++ Content.Server/Radio/EntitySystems/RadioSystem.cs
@@ -22,6 +22,7 @@ using Content.Server.Popups;
 using Content.Server.SS220.Language;
 using System.Diagnostics.CodeAnalysis;
 using Content.Shared.SS220.Language.Systems;  // SS220-Add-Languages
+using Content.Server.SS220.Events; // SS220 borg-id-fix
 
 namespace Content.Server.Radio.EntitySystems;
