+++ Content.Server/SS220/Events/GetInsteadIdCardNameEvent.cs
@@ -1,5 +1,5 @@
 // © SS220, An EULA/CLA with a hosting restriction, full text: https://raw.githubusercontent.com/SerbiaStrong-220/space-station-14/master/CLA.txt
-namespace Content.Server.Radio.EntitySystems;
+namespace Content.Server.SS220.Events;
 
 public sealed class GetInsteadIdCardNameEvent : EntityEventArgs
 {
